+++ CHANGELOG.md
@@ -5,8 +5,10 @@ ISSUES FIXED:
 * [48](https://github.com/perfectsense/gyro-azure-provider/issues/48): Refactor Sql (QA, ResourceFinder, Direct Dependency, Copyable).
 * [49](https://github.com/perfectsense/gyro-azure-provider/issues/49): Public Ip Address Not Set at Startup for Network Interface.
 * [51](https://github.com/perfectsense/gyro-azure-provider/issues/51): Implement Lifecycle for Azure Storage.
+* [55](https://github.com/perfectsense/gyro-azure-provider/issues/55): Azure VirtualMachine should Implement GyroInstance.
 * [56](https://github.com/perfectsense/gyro-azure-provider/issues/56): Implement Azure managed Identities.
 * [62](https://github.com/perfectsense/gyro-azure-provider/issues/62): Add support to manage System Managed Service Identity for VM.
 * [64](https://github.com/perfectsense/gyro-azure-provider/issues/64): Expose ID field for CloudBlobContainerResource and CloudQueueResource.
 * [72](https://github.com/perfectsense/gyro-azure-provider/issues/72): Expose object id for System Managed Identity on Virtual Machine and Scale Set resources.
+* [66](https://github.com/perfectsense/gyro-azure-provider/issues/66): Implement Virtual Machine Image.
 * [70](https://github.com/perfectsense/gyro-azure-provider/issues/70): Add copyright license to java and gradle files.